Space-MMO/Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv
pytest
httpx
alembic
celery[redis]
redis
